+++ src/plugins/index.ts
@@ -2,6 +2,8 @@ import * as amazon from './amazon.js';
 import * as bluesky from './bluesky.js';
 import * as wikipedia from './wikipedia.js';
 import * as branchIoDeeplinks from './branchio-deeplinks.js';
+import * as youtube from './youtube.js';
+import * as spotify from './spotify.js';
 import { SummalyPlugin } from '@/iplugin.js';
 
 export const plugins: SummalyPlugin[] = [
@@ -9,4 +11,6 @@ export const plugins: SummalyPlugin[] = [
 	bluesky,
 	wikipedia,
 	branchIoDeeplinks,
+	youtube,
+	spotify,
 ];